start fleshing out a per-thread setup path for this stuff.
This is going to be separate from the evthr stuff - evthr is fine
as a "take work and throw into queue" mechanism, but it doesn't know
about RSS.

Now, I may end up implementing this in the short term using evthr
stuff - just because it's easy to get it off the ground for what
I need.  But eventually it'll be using external threads and separate
evhtp base / libevent base with a separate listen socket for each.