Start fleshing out a mgr lifecycle.
The aim is to have the manager take care of going through
various run states, waiting for connections to finish,
then deciding when the testing is done or some global
timeout is hit.

The routine in main will then control that and get notified
about the mgr completion; then use this to wrap up things,
print final statistics and exit.