Flesh out skeleton bits for the rest of the states.
The CLEANUP phase doesn't do anything yet - we don't maintain a list
of clt http connections to explicitly go and close.

And right now the cleanup phase is more annoying because we
schedule a destroy of a clt_mgr_conn and it happens in the future.
So in CLEANUP we can call destroy on each, but we need some
way to get notification that they've all destroyed.

I'll go create a new CLEANUP_WAITING phase as part of tracking
the clt_mgr_conn list in clt_mgr.