Overhaul the lifecycle management side of things.
The clt mgr setup stuff is now split into a proper init function,
a config function and a start function.

All the states are doing the right thing now, save cleanup.

I'm not yet doing the actual cleanup - clt mgr conns are now on
a clt mgr list, but I have to add another intermediary state
that waits for cleanup to finish before contiuing to completed.