Fix a double-free thing that's a side effect of how libevhtp does its connection lifecycle management.
If we call evhtp_connection_free() and it has a request with it,
it'll free the request.

If we call evhtp_request_free() first, then it'll free the request,
but the connection its linked to won't be freed.  So, when we then
go and free the connection, it'll double-free the same data and
sadness occurs.