Work around a libevhtp weirdness with how they implement the client-side keepalive code.
If I free the request in clt_req_cb(), it's used via the reference
in evhtp_connection->request, and things go explode.

If I don't free it here, then it doesn't get freed as a keepalive
request.  Only non-keepalive requests are freed - when evhtp_connection_close()
is called, the con->request entry is freed.

So for now, manually disconnect the request from the connection and
free it.  It seems to do the right thing now in both keepalive and
non-keepalive modes.